src/test/java/com/cuongnghiem/springbootrecipe/converters/CategoryCommandToCategoryTest.java
package com.cuongnghiem.springbootrecipe.converters;

import com.cuongnghiem.springbootrecipe.command.CategoryCommand;
import com.cuongnghiem.springbootrecipe.model.Category;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

class CategoryCommandToCategoryTest {

    private final Long ID = 1L;
    private final String DESCRIPTION = "Description";

    CategoryCommandToCategory categoryCommandToCategory;

    @BeforeEach
    void setUp() {
         categoryCommandToCategory = new CategoryCommandToCategory();
    }

    @Test
    void convert() {
        // given
        CategoryCommand categoryCommand = new CategoryCommand();
        categoryCommand.setId(ID);
        categoryCommand.setDescription(DESCRIPTION);

        // when
        Category category = categoryCommandToCategory.convert(categoryCommand);

        // then
        assertNotNull(category);
        assertEquals(ID, category.getId());
        assertEquals(DESCRIPTION, category.getDescription());
    }
}